Fix Docker environment variable substitution in Makefile
The Docker environment variable substitution in the Makefile was not working correctly. This commit fixes the issue by adding the `--env-file` flag to the `docker-compose` command in the `run` target of the Makefile.

Additionally, removed unnecessary code in the `main` function of `cmd/users/main.go`.

This commit addresses the issue and improves the reliability of the Docker environment variable substitution in the project.

Signed-off-by: Rodney Osodo <socials@rodneyosodo.com>